requirements.txt
pandas
ccxt
mysql-connector-python
psycopg2-binary
SQLAlchemy
ta
apache-airflow[celery]
bybit
python-dotenv
